requirements_base.txt
PyQt5==5.13.2
# pyqt5-tools==5.15.2.3
pyrsistent==0.17.3
PySide2==5.15.1
PyYAML==5.3.1
qimage2ndarray==1.8.3
opencv-python==4.4.0.44
numpy==1.19.2
loguru==0.5.3
pandas==1.1.5
tqdm==4.54.1
matplotlib==3.3.3
scipy==1.5.4
yacs==0.1.8
pytest==6.2.4
